+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ### vNEXT
 
+### v0.5.0
+- New [editable field component](./src/field/README.md#editable-field-component) `ign-editable-field`
+- New [editable inline field component](./src/field/README.md#editable-inline-field-component) `ign-editable-field-inline`
+
 ### v0.4.0
 - Introducing first pass of text-list-item component
 - Extracting text-list-item and nav-list into modules